refactor: move components to liquid stream module
