+++ wallet-sdk-ui/src/androidInstrumentedTest/kotlin/com/michaeltchuang/walletsdk/ui/liquidAuth/screens/LiquidAuthOfferScreenshotTest.kt
@@ -14,8 +14,8 @@ import com.michaeltchuang.walletsdk.core.railmpp.core.GatingMode
 import com.michaeltchuang.walletsdk.core.railmpp.core.PaymentRequest
 import com.michaeltchuang.walletsdk.core.railmpp.core.PaymentRequestMeta
 import com.michaeltchuang.walletsdk.ui.base.test.BaseScreenshotTest
-import com.michaeltchuang.walletsdk.ui.liquidAuth.components.CameraStreamingPreviewController
-import com.michaeltchuang.walletsdk.ui.liquidStream.model.IceConnectionType
+import com.michaeltchuang.walletsdk.ui.liquidStream.components.CameraStreamingPreviewController
+import com.michaeltchuang.walletsdk.ui.liquidStream.domain.model.IceConnectionType
 import com.michaeltchuang.walletsdk.ui.liquidAuth.viewmodels.LiquidAuthOfferViewModel
 import org.junit.Test
 import java.util.Locale
