+++ wallet-sdk-ui/src/commonMain/kotlin/com/michaeltchuang/walletsdk/ui/liquidAuth/screens/LiquidAuthOfferScreen.kt
@@ -70,13 +70,13 @@ import com.michaeltchuang.walletsdk.core.railmpp.core.GatingMode
 import com.michaeltchuang.walletsdk.core.railmpp.core.PaymentRequest
 import com.michaeltchuang.walletsdk.core.railmpp.core.PaymentRequestMeta
 import com.michaeltchuang.walletsdk.ui.base.designsystem.theme.AlgoKitTheme
-import com.michaeltchuang.walletsdk.ui.liquidAuth.components.CameraStreamingPreviewController
-import com.michaeltchuang.walletsdk.ui.liquidAuth.components.ConnectedViewersCard
-import com.michaeltchuang.walletsdk.ui.liquidStream.model.IceConnectionType
-import com.michaeltchuang.walletsdk.ui.liquidStream.model.colorHex
-import com.michaeltchuang.walletsdk.ui.liquidStream.model.costTier
-import com.michaeltchuang.walletsdk.ui.liquidStream.model.displayName
-import com.michaeltchuang.walletsdk.ui.liquidStream.model.typicalLatency
+import com.michaeltchuang.walletsdk.ui.liquidStream.components.CameraStreamingPreviewController
+import com.michaeltchuang.walletsdk.ui.liquidStream.components.ConnectedViewersCard
+import com.michaeltchuang.walletsdk.ui.liquidStream.domain.model.IceConnectionType
+import com.michaeltchuang.walletsdk.ui.liquidStream.domain.model.colorHex
+import com.michaeltchuang.walletsdk.ui.liquidStream.domain.model.costTier
+import com.michaeltchuang.walletsdk.ui.liquidStream.domain.model.displayName
+import com.michaeltchuang.walletsdk.ui.liquidStream.domain.model.typicalLatency
 import com.michaeltchuang.walletsdk.ui.liquidAuth.viewmodels.LiquidAuthOfferViewModel
 import com.michaeltchuang.walletsdk.ui.liquidStream.screens.LiquidStreamHostLiveScreen
 import org.jetbrains.compose.ui.tooling.preview.Preview
