+++ wallet-sdk-ui/src/commonMain/kotlin/com/michaeltchuang/walletsdk/ui/liquidStream/screens/LiquidStreamViewerScreen.kt
@@ -53,12 +53,12 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.michaeltchuang.walletsdk.ui.base.designsystem.theme.AlgoKitTheme
-import com.michaeltchuang.walletsdk.ui.liquidAuth.components.ConnectedViewersCard
-import com.michaeltchuang.walletsdk.ui.liquidAuth.components.StreamViewerGiftSupportModal
-import com.michaeltchuang.walletsdk.ui.liquidAuth.components.StreamViewerSettingsSheet
-import com.michaeltchuang.walletsdk.ui.liquidAuth.components.StreamViewerTopUpModel
-import com.michaeltchuang.walletsdk.ui.liquidStream.model.IceConnectionType
-import com.michaeltchuang.walletsdk.ui.liquidStream.model.displayName
+import com.michaeltchuang.walletsdk.ui.liquidStream.components.ConnectedViewersCard
+import com.michaeltchuang.walletsdk.ui.liquidStream.components.StreamViewerGiftSupportModal
+import com.michaeltchuang.walletsdk.ui.liquidStream.components.StreamViewerSettingsSheet
+import com.michaeltchuang.walletsdk.ui.liquidStream.components.StreamViewerTopUpModel
+import com.michaeltchuang.walletsdk.ui.liquidStream.domain.model.IceConnectionType
+import com.michaeltchuang.walletsdk.ui.liquidStream.domain.model.displayName
 import com.michaeltchuang.walletsdk.ui.liquidStream.viewmodels.LiquidAuthViewerViewModel
 import org.jetbrains.compose.resources.DrawableResource
 import org.jetbrains.compose.resources.vectorResource
