+++ wallet-sdk-ui/src/commonMain/kotlin/com/michaeltchuang/walletsdk/ui/liquidStream/screens/StreamHostSettingsSheet.kt
@@ -38,13 +38,13 @@ import com.michaeltchuang.walletsdk.ui.base.designsystem.theme.AlgoKitLightColor
 import com.michaeltchuang.walletsdk.ui.base.designsystem.theme.AlgoKitTheme
 import com.michaeltchuang.walletsdk.ui.base.designsystem.theme.LocalCustomColors
 import com.michaeltchuang.walletsdk.ui.base.designsystem.theme.LocalThemeIsDark
-import com.michaeltchuang.walletsdk.ui.liquidAuth.components.LiquidRingToggle
-import com.michaeltchuang.walletsdk.ui.liquidAuth.components.LiquidSegmentedTabs
-import com.michaeltchuang.walletsdk.ui.liquidAuth.components.SegmentedTabItem
-import com.michaeltchuang.walletsdk.ui.liquidAuth.components.StreamSettingsFootnoteText
-import com.michaeltchuang.walletsdk.ui.liquidAuth.components.StreamSettingsMetricsCard
-import com.michaeltchuang.walletsdk.ui.liquidAuth.components.StreamSettingsPayoutFrequencyBlock
-import com.michaeltchuang.walletsdk.ui.liquidAuth.components.StreamSettingsTopSquareIconButton
+import com.michaeltchuang.walletsdk.ui.liquidStream.components.LiquidRingToggle
+import com.michaeltchuang.walletsdk.ui.liquidStream.components.LiquidSegmentedTabs
+import com.michaeltchuang.walletsdk.ui.liquidStream.components.SegmentedTabItem
+import com.michaeltchuang.walletsdk.ui.liquidStream.components.StreamSettingsFootnoteText
+import com.michaeltchuang.walletsdk.ui.liquidStream.components.StreamSettingsMetricsCard
+import com.michaeltchuang.walletsdk.ui.liquidStream.components.StreamSettingsPayoutFrequencyBlock
+import com.michaeltchuang.walletsdk.ui.liquidStream.components.StreamSettingsTopSquareIconButton
 import com.michaeltchuang.walletsdk.ui.liquidStream.utils.PAYOUT_EVERY_256_BLOCKS_TAB_ID
 import org.jetbrains.compose.ui.tooling.preview.Preview
