+++ wallet-sdk-ui/src/iosMain/kotlin/com/michaeltchuang/walletsdk/ui/liquidStream/components/VideoFrameViewer.ios.kt
@@ -1,4 +1,4 @@
-package com.michaeltchuang.walletsdk.ui.liquidAuth.components
+package com.michaeltchuang.walletsdk.ui.liquidStream.components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
